Move SCM tests to their own class
Too many issues trying to run the test from the TestStorageContainerManager class.
